Author: Dan Nettleton written by David Faden Authors: Rouben Rostamian <rostamian@umbc.edu> Copyright (C) 1999, 2001 the R Development Core Team
in the authors list.

Please add all authors and copyright holders in the Authors@R field with
the appropriate roles.

From the CRAN policies you agreed to:

"The ownership of copyright and intellectual property rights of all
components of the package must be clear and unambiguous (including from
the authors specification in the DESCRIPTION file). Where code is copied
(or derived) from the work of others (including from R itself), care must
be taken that any copyright/license statements are preserved and
authorship is not misrepresented.

Preferably, an ‘Authors@R’ would be used with ‘ctb’ roles for the authors
of such code. Alternatively, the ‘Author’ field should list these authors
as contributors.

Where copyrights are held by an entity other than the package authors,
this should preferably be indicated via ‘cph’ roles in the ‘Authors@R’
field, or using a ‘Copyright’ field (if necessary referring to an
inst/COPYRIGHTS file)."

We also get:
The Date field is not in ISO 8601 yyyy-mm-dd format.

Please add more small executable (outside of \dontrun{}) examples in your
Rd-files.

Your last version has Additional issues:
  rchk
  <https://raw.githubusercontent.com/kalibera/cran-checks/master/rchk/results/EMCluster.out>

  Have these been fixed in this new version?